* src/SharpGit/GitClient/GitStatusEventArgs.h * src/SharpGit/Plumbing/GitTree.cpp * src/SharpGit/Plumbing/GitTree.h   Resolve some warnings by applying casts.

+++ src/SharpGit/GitClient/GitStatusEventArgs.h
@@ -39,7 +39,7 @@ namespace SharpGit {
 			{
 				if (!_relPath && _path)
 				{
-					int n = strlen(_path);
+					int n = (int)strlen(_path);
 
 					if (n > 0 && _path[n-1] == '/')
 						n--;
@@ -57,7 +57,7 @@ namespace SharpGit {
 			{
 				if (!_fullPath && _pool && _path && _wcPath)
 				{
-					int n = strlen(_path);
+					int n = (int)strlen(_path);
 
 					if (n > 0 && _path[n-1] == '/')
 					{
